添加了课程添加业务，修改了课室预约逻辑，同时人脸识别系统可以返回用户人脸图像了

+++ src/main/java/com/postdesign/detectsystem/service/serviceImpl/algorithmImpl/FaceSignSystemServiceImpl.java
@@ -5,21 +5,13 @@ import ai.djl.inference.Predictor;
 import ai.djl.modality.cv.Image;
 import ai.djl.modality.cv.ImageFactory;
 import ai.djl.ndarray.NDArray;
-import ai.djl.ndarray.NDArrays;
-import ai.djl.ndarray.NDManager;
 import ai.djl.repository.zoo.Criteria;
 import ai.djl.repository.zoo.ModelNotFoundException;
 import ai.djl.repository.zoo.ZooModel;
 import ai.djl.training.util.ProgressBar;
-import ai.djl.translate.TranslateException;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.postdesign.detectsystem.configure.WebInterceptor;
-import com.postdesign.detectsystem.entity.Faces;
-import com.postdesign.detectsystem.entity.Student;
-import com.postdesign.detectsystem.entity.TeachCls;
-import com.postdesign.detectsystem.mapper.FaceMapper;
-import com.postdesign.detectsystem.mapper.StudentMapper;
-import com.postdesign.detectsystem.mapper.TeachClsMapper;
+import com.postdesign.detectsystem.entity.*;
+import com.postdesign.detectsystem.mapper.*;
 import com.postdesign.detectsystem.service.algorithmService.FaceSignSystemService;
 import com.postdesign.detectsystem.service.currencyService.RedisService;
 import com.postdesign.detectsystem.translators.CosFaceTranslator;
@@ -32,10 +24,7 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Service
 public class FaceSignSystemServiceImpl implements FaceSignSystemService {
@@ -54,48 +43,45 @@ public class FaceSignSystemServiceImpl implements FaceSignSystemService {
     @Autowired(required = false)
     StudentMapper studentMapper;
 
+    @Autowired(required = false)
+    TeachMajorCourseMapper teachMajorCourseMapper;
+
+    @Autowired(required = false)
+    TeachPublicCourseMapper teachPublicCourseMapper;
+
+    @Autowired(required = false)
+    MajorCourseMapper majorCourseMapper;
+
+    @Autowired(required = false)
+    PublicCourseMapper publicCourseMapper;
+    @Autowired(required = false)
+    ClassMapper classMapper;
+
+    @Autowired(required = false)
+    SignDetailMapper signDetailMapper;
+
+    @Autowired(required = false)
+    UserMapper userMapper;
+
+
     @Override
-    public String importFace(String uid, byte[] face) throws ModelNotFoundException, MalformedModelException, IOException {
-        Faces faces = new Faces();
-        faces.setFaceImg(face);
-        faces.setUid(uid);
-        faceMapper.deleteById(uid);
-        faceMapper.insert(faces);
-        return "success";
-//            Path modelpath = Paths.get("src/main/trained_models/cosface");
-//            ByteArrayInputStream in = new ByteArrayInputStream(face);
-//            Image image = ImageFactory.getInstance().fromInputStream(in);
-//            in.close();
-//            Criteria<Image, NDArray> criteria = Criteria.builder()
-//                    .setTypes(Image.class, NDArray.class)
-//                    .optModelPath(modelpath)
-//                    .optModelName("cosface_cuda.pth")
-//                    .optOption("mapLocation", "true")
-//                    .optTranslator(new CosFaceTranslator())
-//                    .optProgress(new ProgressBar())
-//                    .build();
-//
-//            try (ZooModel<Image, NDArray> zooModel = criteria.loadModel()){
-//                try (Predictor<Image, NDArray> predictor = zooModel.newPredictor()){
-//                    NDArray predict = predictor.predict(image);
-//                    byte[] bytes = predict.toByteArray();
-//                    predict.close();
-//                    predictor.close();
-//                    Faces faces = new Faces();
-//                    faces.setFaceImg(bytes);
-//                    faces.setUid(uid);
-//                    faceMapper.deleteById(uid);
-//                    faceMapper.insert(faces);
-//                    return "success";
-//                } catch (TranslateException e) {
-//                    e.printStackTrace();
-//                    return "error";
-//                }
-//            }
+    public String importFace(String uid, byte[] face) {
+        User user = userMapper.selectById(uid);
+        if (user.getAddfacecount() < 2){
+            Faces faces = new Faces();
+            faces.setFaceImg(face);
+            faces.setUid(uid);
+            faceMapper.deleteById(uid);
+            faceMapper.insert(faces);
+            return "success";
+        }else {
+            return "fail";
+        }
     }
 
     @Override
-    public Map<String, String> signFace(String classname, byte[] face) throws IOException, ModelNotFoundException, MalformedModelException {
+    public Map<String, Object> signFace(String classname, String course, String tid, byte[] face) throws IOException, ModelNotFoundException, MalformedModelException {
+        // 确保返回重新识别时不会重复写入数据
         if (redisService.get(classname) != null){
             redisService.update(classname);
         }else {
@@ -120,27 +106,44 @@ public class FaceSignSystemServiceImpl implements FaceSignSystemService {
         try (ZooModel<Image, NDArray> zooModel = criteria.loadModel()){
             try (Predictor<Image, NDArray> predictor = zooModel.newPredictor()){
                 NDArray detect = predictor.predict(image);
-//                logger.info(detect.toDebugString());
                 String bastUid = null;
                 double bastDis = 0;
+                NDArray predict = null;
+                NDArray distance = null;
                 for (Faces fb:faces){
+                    logger.info("Detecting: " + fb.getUid());
                     ByteArrayInputStream i = new ByteArrayInputStream(fb.getFaceImg());
                     Image f = ImageFactory.getInstance().fromInputStream(i);
                     i.close();
-                    NDArray predict = predictor.predict(f);
-                    NDArray distance = detect.dot(predict.transpose()).div(((detect.norm().mul(predict.norm())).add(1e-5)));
-                    predict.close();
-                    detect.close();
-                    predictor.close();
+                    predict = predictor.predict(f);
+                    distance = detect.dot(predict.transpose()).div(((detect.norm().mul(predict.norm())).add(1e-5)));
                     double[] doubles = distance.toDoubleArray();
-                    distance.close();
                     if (doubles[0] >= bastDis){
                         bastDis = doubles[0];
                         bastUid = fb.getUid();
                     }
                     logger.info(String.valueOf(doubles[0]));
+                    logger.info("Bast: "+bastUid);
                 }
-                if (bastDis >= 0.5){
+                if (bastDis >= 0.4){
+                    User user = userMapper.selectById(bastUid);
+                    byte sign = user.getSign();
+                    if (sign == (byte) 0){
+                        SignDetail signDetail = new SignDetail();
+                        signDetail.setClassname(classname);
+                        signDetail.setCourse(course);
+                        signDetail.setUid(bastUid);
+                        Date date = new Date();
+                        signDetail.setDate(date);
+                        signDetail.setTid(tid);
+                        signDetailMapper.insert(signDetail);
+                        user.setSign((byte) 1);
+                        userMapper.updateById(user);
+                    }
+                    predict.close();
+                    detect.close();
+                    predictor.close();
+                    distance.close();
                     return getBastInfo(bastUid);
                 }else {
                     return null;
@@ -153,7 +156,7 @@ public class FaceSignSystemServiceImpl implements FaceSignSystemService {
     }
 
     @Override
-    public List<String> getSignClasses(String tid) {
+    public Map<String, Object> getSignClasses(String tid) {
         return getTeachClasses(tid);
     }
 
@@ -163,26 +166,69 @@ public class FaceSignSystemServiceImpl implements FaceSignSystemService {
         redisService.set(classname, faceBytes);
     }
 
-    private Map<String, String> getBastInfo(String uid){
+    @Override
+    public Map<String, Byte> getSignDetail(String classname) {
+        QueryWrapper<Student> studentQueryWrapper = new QueryWrapper<>();
+        studentQueryWrapper.eq("cls", classname);
+        List<Student> students = studentMapper.selectList(studentQueryWrapper);
+        Map<String, Byte> map = new HashMap<>();
+        for (Student s:students){
+            map.put(s.getSno()+"\t"+s.getSname(), userMapper.selectById(s.getSno()).getSign());
+        }
+        return map;
+    }
+
+    private Map<String, Object> getBastInfo(String uid){
         if (uid != null){
             Student student = studentMapper.selectById(uid);
-            Map<String, String> detectinfo = new HashMap<>();
+            Faces faces = faceMapper.selectById(uid);
+            Map<String, Object> detectinfo = new HashMap<>();
             detectinfo.put("uno", uid);
             detectinfo.put("name",student.getCls()+" "+student.getSname());
+            detectinfo.put("face", faces.getFaceImg());
             return detectinfo;
         }
         return null;
     }
 
-    private List<String> getTeachClasses(String tid){
+    private Map<String, Object> getTeachClasses(String tid){
+        Map<String, Object> map = new HashMap<>();
+        Map<String, List<String >> m = new HashMap<>();
+        // class
         QueryWrapper<TeachCls> queryWrapper = new QueryWrapper<>();
         queryWrapper.eq("tno", tid);
         List<TeachCls> teachCls = teachClsMapper.selectList(queryWrapper);
         List<String> classes = new ArrayList<>();
         for (TeachCls cls:teachCls){
             classes.add(cls.getClassname());
+            Cls cls1 = classMapper.selectById(cls.getClassname());
+            // course
+            QueryWrapper<TeachMajorCourse> teachMajorCourseQueryWrapper = new QueryWrapper<>();
+            QueryWrapper<TeachPublicCourse> teachPublicCourseQueryWrapper = new QueryWrapper<>();
+            teachMajorCourseQueryWrapper.eq("tno", tid);
+            teachPublicCourseQueryWrapper.eq("tno", tid);
+            List<TeachMajorCourse> teachMajorCourses = teachMajorCourseMapper.selectList(teachMajorCourseQueryWrapper);
+            List<TeachPublicCourse> teachPublicCourses = teachPublicCourseMapper.selectList(teachPublicCourseQueryWrapper);
+            List<String> cs = new ArrayList<>();
+            for (TeachMajorCourse t:teachMajorCourses){
+                MajorCourse majorCourse = majorCourseMapper.selectById(t.getCno());
+                if (majorCourse.getGrade().equals(cls1.getGrade()) && cls1.getMajor().equals(majorCourse.getMajor())){
+                    cs.add(majorCourse.getCname());
+                }
+            }
+
+            for (TeachPublicCourse t:teachPublicCourses){
+                PublicCourse publicCourse = publicCourseMapper.selectById(t.getCpno());
+                if (publicCourse.getGrade().equals(cls1.getGrade())){
+                    cs.add(publicCourse.getCname());
+                }
+            }
+            m.put(cls.getClassname(), cs);
         }
-        return classes;
+        map.put("cls", classes);
+        map.put("cs", m);
+
+        return map;
     }
 
     private List<Faces> getFaceBytes(String classname){
@@ -199,5 +245,15 @@ public class FaceSignSystemServiceImpl implements FaceSignSystemService {
         return facesinfo;
     }
 
+    private void resetSignState(String classname){
+        QueryWrapper<Student> studentQueryWrapper = new QueryWrapper<>();
+        studentQueryWrapper.eq("cls", classname);
+        List<Student> students = studentMapper.selectList(studentQueryWrapper);
+        for (Student s:students){
+            User user = userMapper.selectById(s.getSno());
+            user.setSign((byte) 0);
+            userMapper.updateById(user);
+        }
+    }
 
 }
